latest

+++ contracts/templates/ERC20Enhanced.sol
@@ -3,18 +3,18 @@ pragma solidity ^0.4.23;
 
 contract ERC20Enhanced {
 
-    mapping(address => uint256) balances;
+    mapping(address => uint) balances;
     mapping(address => mapping(address => uint)) internal allowed;
     uint public _totalSupply;
 
-    function allowance(address owner, address spender) public view returns (uint256);
-    function transferFrom(address from, address to, uint256 tokens) public returns (bool);
-    function approve(address spender, uint256 tokens) public returns (bool);
-    function balanceOf(address who) public view returns (uint256);
-    function transfer(address to, uint256 tokens) public returns (bool);
-    function totalSupply() public view returns (uint256);
+    function allowance(address owner, address spender) public view returns (uint);
+    function transferFrom(address from, address to, uint tokens) public returns (bool);
+    function approve(address spender, uint tokens) public returns (bool);
+    function balanceOf(address who) public view returns (uint);
+    function transfer(address to, uint tokens) public returns (bool);
+    function totalSupply() public view returns (uint);
     function migrate(address[] _recipients) public view returns (bool);
-    event Transfer(address indexed from, address indexed to, uint256 value);
-    event Approval(address indexed owner, address indexed spender, uint256 value);
+    event Transfer(address indexed from, address indexed to, uint value);
+    event Approval(address indexed owner, address indexed spender, uint value);
 
 }
